[REFACTOR] 엘라스틱 서치 디렉토리 분리
