+++ src/test/java/spring/hi_hello_spring/HiHelloSpringApplicationTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import spring.hi_hello_spring.config.ElasticsearchIndexInitializer;
-import spring.hi_hello_spring.wiki.query.elasticsearch.repository.WikiDocumentRepository;
+import spring.hi_hello_spring.elasticsearch.command.repository.WikiDocumentRepository;
 
 @SpringBootTest
 class HiHelloSpringApplicationTests {
